made nft card/skelton into components, made explore page dynamic (responsive w/skeleton loading states) , and added filter options to explore page
