+++ src/components/explore/ExploreItems.jsx
@@ -1,75 +1,91 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import AuthorImage from "../../images/author_thumbnail.jpg";
 import nftImage from "../../images/nftImage.jpg";
+import axios from "axios";
+import NFTCardSkeleton from "../UI/NFTSkeletonCard";
+import NFTCard from "../UI/NFTCard";
 
 const ExploreItems = () => {
+  const [loading, setLoading] = useState(false);
+  const [exploreItems, setExploreItems] = useState([]);
+  const [itemsToShow, setItemsToShow] = useState(8);
+
+
+  async function getExploreItems(filter) {
+    setLoading(true);
+    try {
+      if (filter) {
+        const { data } = await axios.get(
+          `https://us-central1-nft-cloud-functions.cloudfunctions.net/explore?filter=${filter}`
+        );
+        setExploreItems(data);
+        setLoading(false);
+      } else {
+        const { data } = await axios.get(
+          "https://us-central1-nft-cloud-functions.cloudfunctions.net/explore"
+        );
+        setExploreItems(data);
+        setLoading(false);
+      }
+    } catch (error) {
+      console.log(error);
+      setLoading(true);
+    }
+  }
+
+  useEffect(() => {
+    getExploreItems("");
+  }, []);
+
+  // remove like function for now, can't get react to update parent from child
+  // const like = (id) => {
+  //   if (likedItems && likedItems.includes(id)) {
+  //     return;
+  //   } else {
+  //     const updateItems = [...exploreItems];
+
+  //     updateItems[id - 1] = {
+  //       ...updateItems[id - 1],
+  //       likes: exploreItems[id - 1].likes + 1,
+  //     };
+  //     setLikedItems([...likedItems, id]);
+  //     setExploreItems(updateItems);
+  //   }
+  // }
+
   return (
     <>
       <div>
-        <select id="filter-items" defaultValue="">
+        <select
+          id="filter-items"
+          defaultValue=""
+          onChange={(e) => getExploreItems(e.target.value)}
+        >
           <option value="">Default</option>
           <option value="price_low_to_high">Price, Low to High</option>
           <option value="price_high_to_low">Price, High to Low</option>
           <option value="likes_high_to_low">Most liked</option>
         </select>
       </div>
-      {new Array(8).fill(0).map((_, index) => (
-        <div
-          key={index}
-          className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
-          style={{ display: "block", backgroundSize: "cover" }}
-        >
-          <div className="nft__item">
-            <div className="author_list_pp">
-              <Link
-                to="/author"
-                data-bs-toggle="tooltip"
-                data-bs-placement="top"
-              >
-                <img className="lazy" src={AuthorImage} alt="" />
-                <i className="fa fa-check"></i>
-              </Link>
-            </div>
-            <div className="de_countdown">5h 30m 32s</div>
+      {loading ? (
+        <NFTCardSkeleton numOfCards={8} />
+      ) : (
+        <NFTCard
+          nftItems={exploreItems.slice(0, itemsToShow)}
+          inSlider={false}
+        />
+      )}
 
-            <div className="nft__item_wrap">
-              <div className="nft__item_extra">
-                <div className="nft__item_buttons">
-                  <button>Buy Now</button>
-                  <div className="nft__item_share">
-                    <h4>Share</h4>
-                    <a href="" target="_blank" rel="noreferrer">
-                      <i className="fa fa-facebook fa-lg"></i>
-                    </a>
-                    <a href="" target="_blank" rel="noreferrer">
-                      <i className="fa fa-twitter fa-lg"></i>
-                    </a>
-                    <a href="">
-                      <i className="fa fa-envelope fa-lg"></i>
-                    </a>
-                  </div>
-                </div>
-              </div>
-              <Link to="/item-details">
-                <img src={nftImage} className="lazy nft__item_preview" alt="" />
-              </Link>
-            </div>
-            <div className="nft__item_info">
-              <Link to="/item-details">
-                <h4>Pinky Ocean</h4>
-              </Link>
-              <div className="nft__item_price">1.74 ETH</div>
-              <div className="nft__item_like">
-                <i className="fa fa-heart"></i>
-                <span>69</span>
-              </div>
-            </div>
-          </div>
-        </div>
-      ))}
-      <div className="col-md-12 text-center">
-        <Link to="" id="loadmore" className="btn-main lead">
+      {/* given */}
+
+      <div className={itemsToShow >= 16 ? "d-none" : "col-md-12 text-center"}>
+        <Link
+          to=""
+          id="loadmore"
+          className="btn-main lead"
+          onClick={() => setItemsToShow(itemsToShow + 4)}
+        >
           Load more
         </Link>
       </div>
