+++ src/pages/Explore.jsx
@@ -1,8 +1,11 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import SubHeader from "../images/subheader.jpg";
 import ExploreItems from "../components/explore/ExploreItems";
+import axios from "axios";
 
 const Explore = () => {
+  
+
   useEffect(() => {
     window.scrollTo(0, 0);
   }, []);
